Move library functionality out of main
