+++ src/lib.rs
@@ -1,9 +1,11 @@
 use std::str;
 
 pub mod graph;
+pub mod imports;
 pub mod migrations;
 pub mod models;
 pub mod test_helpers;
+pub mod updates;
 
 use crate::migrations::run_migrations;
 use noodles::vcf::variant::record::samples::series::value::genotype::Phasing;
